+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v1.2.3
+- Inverted "DontFixMasks" option to "FixMasks" option, which I think is more intuitive
+  - Your config should be updated automatically to use the proper setting
 # v1.2.2
 - Fixed a potential nullref exception with animated particles.
 # v1.2.1